backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
apify-client==1.4.0
google-generativeai==0.3.2
python-dotenv==0.19.0
requests==2.26.0 